requirements.txt
torch==2.8.0.dev20250319+cu128
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.1
langchain_chroma==0.2.5
langchain_ollama==0.3.6
langchain_text_splitters==0.3.9
sentence_transformers==5.1.0
gradio==5.43.1
pynvml==12.0.0
uuid==1.30